fix: fix suggestions (#1)
