+++ contracts/Vault.vy
@@ -869,10 +869,10 @@ def _shareValue(shares: uint256) -> uint256:
     # Determines the current value of `shares`.
         # NOTE: if sqrt(Vault.totalAssets()) >>> 1e39, this could potentially revert
     lockedFundsRatio: uint256 = (block.timestamp - self.lastReport) * self.lockedProfitDegration
-    freeFunds: uint256 = self._totalAssets() 
+    freeFunds: uint256 = self._totalAssets()
     if(lockedFundsRatio < convert(1e18, uint256)):
         freeFunds = freeFunds - (self.lockedProfit - (lockedFundsRatio * self.lockedProfit / convert(1e18, uint256)))
-    
+
     return (shares * freeFunds) / self.totalSupply
